Tidy-up dialects.
